+++ src/main/java/wellink/test/task/repositories/materials/DenimRepository.java
@@ -3,5 +3,5 @@ package wellink.test.task.repositories.materials;
 import org.springframework.data.repository.CrudRepository;
 import wellink.test.task.entities.materials.Denim;
 
-public interface DenimRepository extends CrudRepository <Denim, Integer> {
+public interface DenimRepository extends CrudRepository <Denim, Long> {
 }
