+++ src/main/java/wellink/test/task/repositories/materials/LockRepository.java
@@ -3,5 +3,5 @@ package wellink.test.task.repositories.materials;
 import org.springframework.data.repository.CrudRepository;
 import wellink.test.task.entities.materials.Lock;
 
-public interface LockRepository extends CrudRepository <Lock, Integer> {
+public interface LockRepository extends CrudRepository <Lock, Long> {
 }
