+++ src/main/java/wellink/test/task/repositories/products/DenimJacketRepository.java
@@ -3,5 +3,5 @@ package wellink.test.task.repositories.products;
 import org.springframework.data.repository.CrudRepository;
 import wellink.test.task.entities.products.DenimJacket;
 
-public interface DenimJacketRepository extends CrudRepository <DenimJacket, Integer> {
+public interface DenimJacketRepository extends CrudRepository <DenimJacket, Long> {
 }
