+++ src/main/java/wellink/test/task/repositories/products/DenimPantsRepository.java
@@ -3,5 +3,5 @@ package wellink.test.task.repositories.products;
 import org.springframework.data.repository.CrudRepository;
 import wellink.test.task.entities.products.DenimPants;
 
-public interface DenimPantsRepository extends CrudRepository <DenimPants, Integer> {
+public interface DenimPantsRepository extends CrudRepository <DenimPants, Long> {
 }
