+++ src/main/java/wellink/test/task/repositories/products/DenimSneakersRepository.java
@@ -3,5 +3,5 @@ package wellink.test.task.repositories.products;
 import org.springframework.data.repository.CrudRepository;
 import wellink.test.task.entities.products.DenimSneakers;
 
-public interface DenimSneakersRepository extends CrudRepository <DenimSneakers, Integer> {
+public interface DenimSneakersRepository extends CrudRepository <DenimSneakers, Long> {
 }
